refactor: validateSupporterData, validateDonationData を utils に切り出し + テスト追加
metrics の型ガード関数を services/get-metrics.ts から utils/metrics-validators.ts に
分離し、専用のテストファイルを作成。get-metrics.test.ts からバリデーション関連の
テストを移動。
